refactor: split global cache action factories into resource and infinite files
Split createGlobalCacheActions into createResourceGlobalCacheActions and
createInfiniteQueryGlobalCacheActions, extracting the shared tag predicate
and option types into globalCacheActionsShared. Repoint importers, split
the test file to match, and drop the redundant defaultWrapper in the
useLocalizedKey test (LayerTestProvider is passed directly).
