+++ src/hooks/api/businesses/[business-id]/bank-accounts/useListBankAccounts.ts
@@ -3,8 +3,8 @@ import { Schema } from 'effect'
 import { type BankAccount, BankAccountSchema } from '@schemas/bankAccounts/bankAccount'
 import { UnwrappedDataResponseSchema } from '@schemas/utils'
 import { get } from '@utils/api/authenticatedHttp'
-import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createGlobalCacheActions'
 import { createQueryHook } from '@hooks/utils/swr/createQueryHook'
+import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createResourceGlobalCacheActions'
 
 export const BANK_ACCOUNTS_TAG_KEY = '#bank-accounts'
