+++ src/hooks/api/businesses/[business-id]/custom-accounts/useCustomAccounts.ts
@@ -3,8 +3,8 @@ import { pipe, Schema } from 'effect'
 import { CustomAccountSchema } from '@schemas/customAccounts'
 import { UnwrappedDataResponseSchema } from '@schemas/utils'
 import { getWithQuery } from '@utils/api/getWithQuery'
-import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createGlobalCacheActions'
 import { createQueryHook } from '@hooks/utils/swr/createQueryHook'
+import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createResourceGlobalCacheActions'
 
 export const CUSTOM_ACCOUNTS_TAG_KEY = '#custom-accounts'
