+++ src/hooks/api/businesses/[business-id]/invoices/useListInvoices.tsx
@@ -2,7 +2,7 @@ import { type PaginationParams, SortOrder, type SortParams } from '@internal-typ
 import { PaginatedResponseSchema } from '@schemas/common/pagination'
 import { type Invoice, InvoiceSchema, type InvoiceStatus } from '@schemas/invoices/invoice'
 import { getWithQuery } from '@utils/api/getWithQuery'
-import { createInfiniteQueryGlobalCacheActions } from '@hooks/utils/swr/createGlobalCacheActions'
+import { createInfiniteQueryGlobalCacheActions } from '@hooks/utils/swr/createInfiniteQueryGlobalCacheActions'
 import { createInfiniteQueryHook } from '@hooks/utils/swr/createInfiniteQueryHook'
 
 export const LIST_INVOICES_TAG_KEY = '#list-invoices'
