+++ src/hooks/api/businesses/[business-id]/mileage/summary/useMileageSummary.tsx
@@ -1,8 +1,8 @@
 import { type MileageSummary, MileageSummarySchema } from '@schemas/mileage'
 import { UnwrappedDataResponseSchema } from '@schemas/utils'
 import { get } from '@utils/api/authenticatedHttp'
-import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createGlobalCacheActions'
 import { createQueryHook } from '@hooks/utils/swr/createQueryHook'
+import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createResourceGlobalCacheActions'
 
 export const MILEAGE_SUMMARY_TAG_KEY = '#mileage-summary'
