+++ src/hooks/api/businesses/[business-id]/mileage/vehicles/useListVehicles.ts
@@ -4,8 +4,8 @@ import { Schema } from 'effect'
 import { UnwrappedDataResponseSchema } from '@schemas/utils'
 import { type Vehicle, VehicleSchema } from '@schemas/vehicle'
 import { getWithQuery } from '@utils/api/getWithQuery'
-import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createGlobalCacheActions'
 import { createQueryHook } from '@hooks/utils/swr/createQueryHook'
+import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createResourceGlobalCacheActions'
 
 type ListVehiclesParams = {
   businessId: string
