+++ src/hooks/api/businesses/[business-id]/tax-estimates/banner/useTaxEstimatesBanner.ts
@@ -1,8 +1,8 @@
 import { type TaxEstimatesBanner, TaxEstimatesBannerResponseSchema } from '@schemas/taxEstimates/banner'
 import { getWithQuery } from '@utils/api/getWithQuery'
 import { type TaxEstimatesRequestParams, toTaxEstimatesQuery } from '@hooks/api/businesses/[business-id]/tax-estimates/taxEstimatesParams'
-import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createGlobalCacheActions'
 import { createQueryHook } from '@hooks/utils/swr/createQueryHook'
+import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createResourceGlobalCacheActions'
 
 const TAX_ESTIMATES_BANNER_TAG_KEY = '#tax-estimates-banner'
