+++ src/hooks/api/businesses/[business-id]/tax-estimates/details/useTaxDetails.ts
@@ -1,8 +1,8 @@
 import { type TaxDetails, TaxDetailsResponseSchema } from '@schemas/taxEstimates/details'
 import { getWithQuery } from '@utils/api/getWithQuery'
 import { type TaxEstimatesRequestParams, toTaxEstimatesQuery } from '@hooks/api/businesses/[business-id]/tax-estimates/taxEstimatesParams'
-import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createGlobalCacheActions'
 import { createQueryHook } from '@hooks/utils/swr/createQueryHook'
+import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createResourceGlobalCacheActions'
 
 const TAX_DETAILS_TAG_KEY = '#tax-details'
