+++ src/hooks/api/businesses/[business-id]/tax-estimates/payments/useTaxPayments.ts
@@ -1,8 +1,8 @@
 import { type TaxPaymentRow, TaxPaymentsResponseSchema } from '@schemas/taxEstimates/payments'
 import { getWithQuery } from '@utils/api/getWithQuery'
 import { type TaxEstimatesRequestParams, toTaxEstimatesQuery } from '@hooks/api/businesses/[business-id]/tax-estimates/taxEstimatesParams'
-import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createGlobalCacheActions'
 import { createQueryHook } from '@hooks/utils/swr/createQueryHook'
+import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createResourceGlobalCacheActions'
 
 const TAX_PAYMENTS_TAG_KEY = '#tax-payments'
