+++ src/hooks/api/businesses/[business-id]/tax-estimates/summary/useTaxSummary.ts
@@ -1,8 +1,8 @@
 import { type TaxSummary, TaxSummaryResponseSchema } from '@schemas/taxEstimates/summary'
 import { getWithQuery } from '@utils/api/getWithQuery'
 import { type TaxEstimatesRequestParams, toTaxEstimatesQuery } from '@hooks/api/businesses/[business-id]/tax-estimates/taxEstimatesParams'
-import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createGlobalCacheActions'
 import { createQueryHook } from '@hooks/utils/swr/createQueryHook'
+import { createResourceGlobalCacheActions } from '@hooks/utils/swr/createResourceGlobalCacheActions'
 
 const TAX_SUMMARY_TAG_KEY = '#tax-summary'
